🚚 Rename Error folder => "error"
