+++ packages/apps/web-api/src/error/BadRequestError.ts
